Refactor: separate CLI and API layers
- Move CLI logic from core/calculator.py to new cli/ directory
- Extract core functions to api.py for library usage
- Update entry point from core.calculator to cli.main
- Add AGENTS.md to .gitignore
- Improve console width handling with MIN_CONSOLE_WIDTH
- Update test imports to use new CLI structure

This refactoring separates concerns between CLI interface and
programmatic API, making the library easier to use as a dependency.
