+++ setup.py
@@ -46,7 +46,7 @@ setup(
     },
     entry_points={
         "console_scripts": [
-            "vt-calc=vt_calculator.core.calculator:main",
+            "vt-calc=vt_calculator.cli.main:main",
         ],
     },
     keywords=[
