+++ src/vt_calculator/__init__.py
@@ -8,13 +8,26 @@ with various Vision Language Models (VLMs).
 __version__ = "0.0.3"
 __author__ = "Vision Token Calculator"
 
-from .core.calculator import count_image_tokens, process_directory
-from .utils import create_dummy_image
-from .reporter import display_batch_results
+from .api import (
+    count_image_tokens,
+    count_video_tokens,
+    process_directory,
+    parse_compare_models,
+    compare_image_tokens,
+    compare_video_tokens,
+)
+from .utils import create_dummy_image, create_dummy_video
+from .reporter import display_batch_results, display_comparison_results
 
 __all__ = [
     "count_image_tokens",
-    "create_dummy_image",
+    "count_video_tokens",
     "process_directory",
+    "parse_compare_models",
+    "compare_image_tokens",
+    "compare_video_tokens",
+    "create_dummy_image",
+    "create_dummy_video",
     "display_batch_results",
+    "display_comparison_results",
 ]
