+++ src/vt_calculator/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI module for vt-calc command."""
+
+from .main import main
+
+__all__ = ["main"]
